Added new simulations
Added new simulations: 3D_ver_2 is for a higher resolution, 3D_obstacle is with a long plane where there's a pillar to simulate the disruption of the flow.

Note that the Python for the obstacle render is slow. It is better to rewrite it in C++ for faster renders, since there's 1.4 billion (Grid points * frames = 150 * 75 * 150 * 800 = 1,350,000,000) calculations needed.